+++ src/main/java/ba/atlantbh/auctionapp/responses/ProductResponse.java
@@ -4,6 +4,7 @@ import ba.atlantbh.auctionapp.models.Photo;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -15,7 +16,7 @@ public class ProductResponse {
     private UUID personId;
     private String name;
     private String description;
-    private Float startPrice;
+    private BigDecimal startPrice;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
     private Boolean wished;
