+++ src/main/java/ba/atlantbh/auctionapp/responses/SimpleProductResponse.java
@@ -1,11 +1,12 @@
 package ba.atlantbh.auctionapp.responses;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public interface SimpleProductResponse {
     UUID getId();
     String getName();
-    Float getStartPrice();
+    BigDecimal getStartPrice();
     String getDescription();
     String getUrl();
     String getCategoryName();
